backend/requirements.txt
mysql-connector-python
python-dotenv
PyMySQL
azure-functions
aniso8601
click
Flask
Flask-RESTful
Flask-SQLAlchemy
itsdangerous
Jinja2
MarkupSafe
pytz
pylint-flask
pylint
pylint_flask_sqlalchemy
six
SQLAlchemy
Werkzeug
flask-cors